refactor: hook 분리
